+++ lib/countly_flutter.dart
@@ -58,7 +58,6 @@ class Countly {
     _userProfileInternal = UserProfileInternal(this, _countlyState);
     _viewsInternal = ViewsInternal(this, _countlyState);
     _sessionsInternal = SessionsInternal(this, _countlyState);
-
   }
   static final instance = _instance;
   static final _instance = Countly._();
@@ -239,7 +238,7 @@ class Countly {
     }
     if (config.manualSessionEnabled != null) {
       _manualSessionControlEnabled = config.manualSessionEnabled!;
-      if(config.manualSessionEnabled!) {
+      if (config.manualSessionEnabled!) {
         _instance._sessionsInternal.enableManualSession();
       }
     }
@@ -579,7 +578,6 @@ class Countly {
 
     return result;
   }
-
 
   @Deprecated('Automatic sessions are handled by underlying SDK, this function will do nothing')
   static Future<String?> start() async {
@@ -2003,8 +2001,12 @@ class Countly {
         countlyConfig['globalViewSegmentation'] = config.globalViewSegmentation;
       }
 
-      if (config.enableAllConsents != null) {
+      if (config.enableAllConsents) {
         countlyConfig['enableAllConsents'] = config.enableAllConsents;
+      }
+
+      if (config.autoEnrollABOnDownload) {
+        countlyConfig['autoEnrollABOnDownload'] = config.autoEnrollABOnDownload;
       }
 
       countlyConfig['remoteConfigAutomaticTriggers'] = config.remoteConfigAutomaticTriggers;