+++ ios/Classes/CountlyiOS/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 23.8.2
+- Fixed rating feedback widget event key for widget closed event
+- Added `testingDownloadExperimentInformation:` in remote config interface
+- Added `testingGetAllExperimentInfo:` in remote config interface
+
+## 23.8.1
+- Expanded feedback widget functionality. Added ability to use rating widgets.
+- Added functionality to access tags for feedback widgets.
+- Fixed SPM public header issues of `CountlyViewTracking.h` 
+
 ## 23.8.0
 - Added `CountlyViewTracking:` interface with new view methods:
     - `setGlobalViewSegmentation:`